feat: devnet overall stable
